[GH-1078]: renamed the loading file to loading/index.tsx
